+++ src/sections/TimeSheet/components/DateSort/index.js
@@ -6,9 +6,9 @@ import type { Dispatch } from 'redux';
 
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 
-import { changeDateSort } from '../../actions/time';
+import { changeDateSort } from '../../../../actions/time';
 
-import { getDateSort } from '../../selectors/time';
+import { getDateSort } from '../../../../selectors/time';
 
 type DateSortType = {
   sort: sortDirection;
